dispatch: make assertions LLP64 friendly (NFCI)
Use `long long` rather than `long` for the value being asserted.  This
is important as there are many places where we use the assertions on
pointer values, which would be truncated on a LLP64 environment.  This
helps clean up the warning spew on Windows.

+++ src/internal.h
@@ -587,11 +587,11 @@ void _dispatch_log(const char *msg, ...);
  */
 DISPATCH_ALWAYS_INLINE
 static inline void
-_dispatch_assert(long e, size_t line) DISPATCH_STATIC_ASSERT_IF(!e)
+_dispatch_assert(long long e, size_t line) DISPATCH_STATIC_ASSERT_IF(!e)
 {
 	if (unlikely(DISPATCH_DEBUG && !e)) _dispatch_abort(line, e);
 }
-#define dispatch_assert(e) _dispatch_assert((long)(e), __LINE__)
+#define dispatch_assert(e) _dispatch_assert((long long)(e), __LINE__)
 
 /*
  * A lot of API return zero upon success and not-zero on fail. Let's capture
@@ -599,11 +599,11 @@ _dispatch_assert(long e, size_t line) DISPATCH_STATIC_ASSERT_IF(!e)
  */
 DISPATCH_ALWAYS_INLINE
 static inline void
-_dispatch_assert_zero(long e, size_t line) DISPATCH_STATIC_ASSERT_IF(e)
+_dispatch_assert_zero(long long e, size_t line) DISPATCH_STATIC_ASSERT_IF(e)
 {
 	if (unlikely(DISPATCH_DEBUG && e)) _dispatch_abort(line, e);
 }
-#define dispatch_assert_zero(e) _dispatch_assert_zero((long)(e), __LINE__)
+#define dispatch_assert_zero(e) _dispatch_assert_zero((long long)(e), __LINE__)
 
 /*
  * For reporting bugs or impedance mismatches between libdispatch and external
@@ -613,12 +613,12 @@ _dispatch_assert_zero(long e, size_t line) DISPATCH_STATIC_ASSERT_IF(e)
  */
 DISPATCH_ALWAYS_INLINE
 static inline void
-_dispatch_assume(long e, size_t line) DISPATCH_STATIC_ASSERT_IF(!e)
+_dispatch_assume(long long e, size_t line) DISPATCH_STATIC_ASSERT_IF(!e)
 {
 	if (unlikely(!e)) _dispatch_bug(line, e);
 }
 #define dispatch_assume(e) \
-		({ __typeof__(e) _e = (e); _dispatch_assume((long)_e, __LINE__); _e; })
+		({ __typeof__(e) _e = (e); _dispatch_assume((long long)_e, __LINE__); _e; })
 
 /*
  * A lot of API return zero upon success and not-zero on fail. Let's capture
